obs-clock-tool/requirements.txt
PyQt5==5.15.9
obs-websocket-py==1.0.0
python-dateutil==2.8.2
pytz==2024.1 